extract FrameBitmapCache and FrameBitmapStream base from ImaVidImageStream
